Create new app profile and refactoring user app
